Fix FOV mixin return type crash on 1.20.5-1.21.1 (bump 1.2.1)
GameRenderer.getFov returns double up to 1.21.1 and float from 1.21.2.
The 1.20.5/1.20.6/1.21/1.21.1 modules declared a Float CallbackInfoReturnable,
so reading the captured return value threw ClassCastException
(Double -> Float) as soon as a world rendered - e.g. right after joining
a server. Those 8 modules now use a Double callback like all other
double-returning versions (verified against the official Mojang mappings
for every version and against the 26.x jars for the Camera hook).
Mod version bumped to 1.2.1 so the fixed jars replace 1.2.0 on Modrinth
(existing version numbers are skipped by the sync).

+++ versions/neoforge-1_20_6/src/main/java/com/vzoom/client/mixin/GameRendererMixin.java
@@ -27,7 +27,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public class GameRendererMixin {
 
     @Inject(method = "getFov", at = @At("RETURN"), cancellable = true)
-    private void vzoom$modifyFov(CallbackInfoReturnable<Float> cir) {
+    private void vzoom$modifyFov(CallbackInfoReturnable<Double> cir) {
         ZoomState s = ZoomState.INSTANCE;
         // Tick here rather than from the HUD callback: the world (and this
         // FOV hook) still runs while the HUD is hidden with F1.
@@ -35,6 +35,6 @@ public class GameRendererMixin {
                 com.vzoom.client.VZoomClient.isStateZoomable());
         Minecraft mc = Minecraft.getInstance();
         mc.options.smoothCamera = s.isActive() && s.config.cinematicCamera;
-        cir.setReturnValue((float) s.computeFov(cir.getReturnValue()));
+        cir.setReturnValue(s.computeFov(cir.getReturnValue()));
     }
 }
